Elevator floors Figure finished Dressable finished

+++ trainging_for_mirela/src/main/java/figure/Figure.java
@@ -1,41 +1,7 @@
 package figure;
 
 public abstract class Figure {
-
-//    private int height, width, length, radius;
-//
-//    public int getHeight() {
-//        return height;
-//    }
-//
-//    public void setHeight(int height) {
-//        this.height = height;
-//    }
-//
-//    public int getLength() {
-//        return length;
-//    }
-//
-//    public void setLength(int length) {
-//        this.length = length;
-//    }
-//
-//    public int getRadius() {
-//        return radius;
-//    }
-//
-//    public void setRadius(int radius) {
-//        this.radius = radius;
-//    }
-//
-//    public int getWidth() {
-//        return width;
-//    }
-//
-//    public void setWidth(int width) {
-//        this.width = width;
-//    }
-    public abstract int getPerimeter();
+    public abstract double getPerimeter();
 
     public abstract double getArea();
 }
